+++ update/org.eclipse.update.core/src/org/eclipse/update/core/IPluginEntry.java
@@ -99,7 +99,7 @@ public interface IPluginEntry {
 	 * @since 2.0 
 	 */
 
-	int getDownloadSize();
+	long getDownloadSize();
 
 
 	/**
@@ -112,7 +112,7 @@ public interface IPluginEntry {
 	 * @since 2.0 
 	 */
 
-	int getInstallSize(); 
+	long getInstallSize(); 
 	
 }
